Moved scoring functionality to QRCodeUtil, added QRCodeUtilTest class
