+++ QRHunter/app/src/main/java/com/example/qrhunter/data/repository/QRCodeRepository.java
@@ -1,50 +1,13 @@
 package com.example.qrhunter.data.repository;
 
-import androidx.annotation.NonNull;
+import com.example.qrhunter.data.model.QRCode;
 
-import org.apache.commons.codec.digest.DigestUtils;
-
-public class QRCodeRepository {
-    public String hashQR (String qrCode) {
-        return DigestUtils.sha256Hex(qrCode);
+public class QRCodeRepository extends DataRepository {
+    /**
+     * Create a QR Code document in Firestore
+     */
+    public void addQRCode(QRCode qrCode) {
+        // Add QR Code to firestore
+        // db.document().set()
     }
-    public double calculateScore(@NonNull String hash) {
-        double score = 0;
-        char[] array = hash.toCharArray();
-        for (int i = 0; i < array.length-1; i++) {
-            int value = 0;
-            if (array[i] == '0' && array[i+1] != '0') {
-                score += 1;
-            }
-            else if (array[i] == array [i+1] ) {
-                int j = i + 1;
-                while (array[i] == array[j]) {
-                    value += 1;
-                    j += 1;
-                    if (j == array.length) {
-                        break;
-                    }
-                }
-            }
-            if (value > 0) {
-                score += calculatePoints(array[i], value);
-                i += value;
-            }
-        }
-        return score;
-    }
-
-    public double calculatePoints(Character key, Integer value) {
-        double points;
-        int ascii = (int) key;
-        if (ascii < 58) {
-            ascii -= 48;
-        }
-        else {
-            ascii -= 87;
-        }
-        points = Math.pow(ascii, value);
-        return points;
-    }
-
 }
